+++ MatrixKit/Models/Search/MXKSearchDataSource.m
@@ -120,23 +120,33 @@ NSString *const kMXKSearchCellDataIdentifier = @"kMXKSearchCellDataIdentifier";
     return cellData;
 }
 
-- (void)convertHomeserverResultsIntoCells:(MXSearchRoomEventResults*)roomEventResults
+- (void)convertHomeserverResultsIntoCells:(MXSearchRoomEventResults*)roomEventResults onComplete:(dispatch_block_t)onComplete
 {
     // Retrieve the MXKCellData class to manage the data
     // Note: MXKSearchDataSource only manages MXKCellData that conforms to MXKSearchCellDataStoring protocol
     // see `[registerCellDataClass:forCellIdentifier:]`
     Class class = [self cellDataClassForCellIdentifier:kMXKSearchCellDataIdentifier];
 
+    dispatch_group_t group = dispatch_group_create();
+
     for (MXSearchResult *result in roomEventResults.results)
     {
-        MXKCellData  *cellData = [[class alloc] initWithSearchResult:result andSearchDataSource:self];
-        if (cellData)
-        {
-            ((id<MXKSearchCellDataStoring>)cellData).shouldShowRoomDisplayName = self.shouldShowRoomDisplayName;
-            
-            [cellDataArray insertObject:cellData atIndex:0];
-        }
+        dispatch_group_enter(group);
+        [[class alloc] cellDataWithSearchResult:result andSearchDataSource:self onComplete:^(__autoreleasing id<MXKSearchCellDataStoring> cellData) {
+            dispatch_group_leave(group);
+
+            if (cellData)
+            {
+                ((id<MXKSearchCellDataStoring>)cellData).shouldShowRoomDisplayName = self.shouldShowRoomDisplayName;
+
+                [self->cellDataArray insertObject:cellData atIndex:0];
+            }
+        }];
     }
+
+    dispatch_group_notify(group, dispatch_get_main_queue(), ^{
+        onComplete();
+    });
 }
 
 #pragma mark - Private methods
@@ -175,18 +185,18 @@ NSString *const kMXKSearchCellDataIdentifier = @"kMXKSearchCellDataIdentifier";
         _canPaginate = (nil != nextBatch);
 
         // Process HS response to cells data
-        [self convertHomeserverResultsIntoCells:roomEventResults];
+        [self convertHomeserverResultsIntoCells:roomEventResults onComplete:^{
+            self.state = MXKDataSourceStateReady;
 
-        self.state = MXKDataSourceStateReady;
+            // Provide changes information to the delegate
+            NSIndexSet *insertedIndexes;
+            if (roomEventResults.results.count)
+            {
+                insertedIndexes = [NSIndexSet indexSetWithIndexesInRange:NSMakeRange(0, roomEventResults.results.count)];
+            }
 
-        // Provide changes information to the delegate
-        NSIndexSet *insertedIndexes;
-        if (roomEventResults.results.count)
-        {
-            insertedIndexes = [NSIndexSet indexSetWithIndexesInRange:NSMakeRange(0, roomEventResults.results.count)];
-        }
-
-        [self.delegate dataSource:self didCellChange:insertedIndexes];
+            [self.delegate dataSource:self didCellChange:insertedIndexes];
+        }];
 
     } failure:^(NSError *error) {
         searchRequest = nil;