+++ MatrixKit/Controllers/MXKRoomViewController.m
@@ -49,18 +49,7 @@
 
 #import "NSBundle+MatrixKit.h"
 #import "UIScrollView+MatrixKit.h"
-
-NSString *const kCmdChangeDisplayName = @"/nick";
-NSString *const kCmdEmote = @"/me";
-NSString *const kCmdJoinRoom = @"/join";
-NSString *const kCmdPartRoom = @"/part";
-NSString *const kCmdInviteUser = @"/invite";
-NSString *const kCmdKickUser = @"/kick";
-NSString *const kCmdBanUser = @"/ban";
-NSString *const kCmdUnbanUser = @"/unban";
-NSString *const kCmdSetUserPowerLevel = @"/op";
-NSString *const kCmdResetUserPowerLevel = @"/deop";
-NSString *const kCmdChangeRoomTopic = @"/topic";
+#import "MXKSlashCommands.h"
 
 @interface MXKRoomViewController ()
 {
@@ -1313,20 +1302,20 @@ NSString *const kCmdChangeRoomTopic = @"/topic";
     // TODO: display an alert with the cmd usage in case of error or unrecognized cmd.
     NSString *cmdUsage;
     
-    if ([cmd isEqualToString:kCmdEmote])
+    if ([cmd isEqualToString:kMXKSlashCmdEmote])
     {
         // send message as an emote
         [self sendTextMessage:string];
     }
-    else if ([string hasPrefix:kCmdChangeDisplayName])
+    else if ([string hasPrefix:kMXKSlashCmdChangeDisplayName])
     {
         // Change display name
         NSString *displayName;
         
         // Sanity check
-        if (string.length > kCmdChangeDisplayName.length)
+        if (string.length > kMXKSlashCmdChangeDisplayName.length)
         {
-            displayName = [string substringFromIndex:kCmdChangeDisplayName.length + 1];
+            displayName = [string substringFromIndex:kMXKSlashCmdChangeDisplayName.length + 1];
             
             // Remove white space from both ends
             displayName = [displayName stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceCharacterSet]];
@@ -1351,15 +1340,15 @@ NSString *const kCmdChangeRoomTopic = @"/topic";
             cmdUsage = @"Usage: /nick <display_name>";
         }
     }
-    else if ([string hasPrefix:kCmdJoinRoom])
+    else if ([string hasPrefix:kMXKSlashCmdJoinRoom])
     {
         // Join a room
         NSString *roomAlias;
         
         // Sanity check
-        if (string.length > kCmdJoinRoom.length)
+        if (string.length > kMXKSlashCmdJoinRoom.length)
         {
-            roomAlias = [string substringFromIndex:kCmdJoinRoom.length + 1];
+            roomAlias = [string substringFromIndex:kMXKSlashCmdJoinRoom.length + 1];
             
             // Remove white space from both ends
             roomAlias = [roomAlias stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceCharacterSet]];
@@ -1385,16 +1374,16 @@ NSString *const kCmdChangeRoomTopic = @"/topic";
             cmdUsage = @"Usage: /join <room_alias>";
         }
     }
-    else if ([string hasPrefix:kCmdPartRoom])
+    else if ([string hasPrefix:kMXKSlashCmdPartRoom])
     {
         // Leave this room or another one
         NSString *roomId;
         NSString *roomIdOrAlias;
         
         // Sanity check
-        if (string.length > kCmdPartRoom.length)
+        if (string.length > kMXKSlashCmdPartRoom.length)
         {
-            roomIdOrAlias = [string substringFromIndex:kCmdPartRoom.length + 1];
+            roomIdOrAlias = [string substringFromIndex:kMXKSlashCmdPartRoom.length + 1];
             
             // Remove white space from both ends
             roomIdOrAlias = [roomIdOrAlias stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceCharacterSet]];
@@ -1443,15 +1432,15 @@ NSString *const kCmdChangeRoomTopic = @"/topic";
             cmdUsage = @"Usage: /part [<room_alias>]";
         }
     }
-    else if ([string hasPrefix:kCmdChangeRoomTopic])
+    else if ([string hasPrefix:kMXKSlashCmdChangeRoomTopic])
     {
         // Change topic
         NSString *topic;
         
         // Sanity check
-        if (string.length > kCmdChangeRoomTopic.length)
+        if (string.length > kMXKSlashCmdChangeRoomTopic.length)
         {
-            topic = [string substringFromIndex:kCmdChangeRoomTopic.length + 1];
+            topic = [string substringFromIndex:kMXKSlashCmdChangeRoomTopic.length + 1];
             // Remove white space from both ends
             topic = [topic stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceCharacterSet]];
         }
@@ -1491,7 +1480,7 @@ NSString *const kCmdChangeRoomTopic = @"/topic";
             userId = nil;
         }
         
-        if ([cmd isEqualToString:kCmdInviteUser])
+        if ([cmd isEqualToString:kMXKSlashCmdInviteUser])
         {
             if (userId)
             {
@@ -1513,7 +1502,7 @@ NSString *const kCmdChangeRoomTopic = @"/topic";
                 cmdUsage = @"Usage: /invite <userId>";
             }
         }
-        else if ([cmd isEqualToString:kCmdKickUser])
+        else if ([cmd isEqualToString:kMXKSlashCmdKickUser])
         {
             if (userId)
             {
@@ -1548,7 +1537,7 @@ NSString *const kCmdChangeRoomTopic = @"/topic";
                 cmdUsage = @"Usage: /kick <userId> [<reason>]";
             }
         }
-        else if ([cmd isEqualToString:kCmdBanUser])
+        else if ([cmd isEqualToString:kMXKSlashCmdBanUser])
         {
             if (userId)
             {
@@ -1583,7 +1572,7 @@ NSString *const kCmdChangeRoomTopic = @"/topic";
                 cmdUsage = @"Usage: /ban <userId> [<reason>]";
             }
         }
-        else if ([cmd isEqualToString:kCmdUnbanUser])
+        else if ([cmd isEqualToString:kMXKSlashCmdUnbanUser])
         {
             if (userId)
             {
@@ -1605,7 +1594,7 @@ NSString *const kCmdChangeRoomTopic = @"/topic";
                 cmdUsage = @"Usage: /unban <userId>";
             }
         }
-        else if ([cmd isEqualToString:kCmdSetUserPowerLevel])
+        else if ([cmd isEqualToString:kMXKSlashCmdSetUserPowerLevel])
         {
             // Retrieve power level
             NSString *powerLevel = nil;
@@ -1641,7 +1630,7 @@ NSString *const kCmdChangeRoomTopic = @"/topic";
                 cmdUsage = @"Usage: /op <userId> <power level>";
             }
         }
-        else if ([cmd isEqualToString:kCmdResetUserPowerLevel])
+        else if ([cmd isEqualToString:kMXKSlashCmdResetUserPowerLevel])
         {
             if (userId)
             {