+++ CHANGES.rst
@@ -7,6 +7,7 @@ Bug fix:
 
 API break:
  * MXKContactManager: Remove the privateMatrixContacts method.
+ * MXKSearchCellDataStoring: Replace initWithSearchResult by async cellDataWithSearchResult.
 
 Changes in MatrixKit in 0.7.15 (2018-07-03)
 ==========================================