+++ CHANGES.rst
@@ -1,3 +1,10 @@
+Changes in MatrixKit in 0.7.16 ()
+==========================================
+
+Improvements:
+ * MXKRoomDataSource: Add send reply with text message (vector-im/riot-ios#1911).
+
+
 Changes in MatrixKit in 0.7.15 (2018-07-03)
 ==========================================